Create notif.py
This code includes the get_current_price() function, which uses an API call to retrieve the current price of the given token, and the send_notification() function, which sends a notification with the given message.

The check_price() function uses these helper functions to check the current price of the given token and send a notification if it is above or below the given threshold. It also includes an additional else clause to handle the case where the current price is equal to the threshold.

To use this code in a real-world application, you will need to replace the API call in the get_current_price() function with a real API call that retrieves the current price of the token, and you will need to replace the print() statement in the send_notification() function with code to send a real notification.